E Commerce Backend/requirements.txt
open_clip_torch
torchvision
pillow
pymssql
numpy
google-generativeai